- [x] Created ElementCompleteService Trait 	- [x] Refactored ElementCompleteService 	- [x] Refactored OrganizationElementCompleteService
