+++ app/IATI/Services/ElementCompleteService.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\IATI\Services;
 
+use App\IATI\Traits\ElementCompleteServiceTrait;
 use Illuminate\Support\Arr;
 
 /**
@@ -11,6 +12,7 @@ use Illuminate\Support\Arr;
  */
 class ElementCompleteService
 {
+    use ElementCompleteServiceTrait;
     /**
      * Public variable element.
      *
@@ -27,103 +29,6 @@ class ElementCompleteService
      * @var string
      */
     public $tempAmount = '';
-
-    /**
-     * Returns mandatory fields.
-     *
-     * @param $field
-     *
-     * @return bool
-     */
-    public function isFieldMandatory($field): bool
-    {
-        return array_key_exists('criteria', $field) && $field['criteria'] === 'mandatory';
-    }
-
-    /**
-     * @param       $fields
-     * @param array $mandatoryFields
-     *
-     * @return array
-     */
-    public function getMandatoryFields($fields, array $mandatoryFields): array
-    {
-        foreach ($fields as $attribute) {
-            if ($this->isFieldMandatory($attribute)) {
-                $mandatoryFields[] = $attribute['name'];
-            }
-        }
-
-        return $mandatoryFields;
-    }
-
-    /**
-     * Return mandatory attributes fields.
-     *
-     * @param $attributes
-     *
-     * @return array
-     */
-    public function mandatoryAttributes($attributes): array
-    {
-        return $this->getMandatoryFields($attributes, []);
-    }
-
-    /**
-     * Return mandatory sub element fields.
-     *
-     * @param $fields
-     *
-     * @return array
-     */
-    public function mandatorySubElements($fields): array
-    {
-        $mandatoryElements = [];
-
-        foreach ($fields as $field) {
-            $mandatoryFields = [];
-
-            if ($this->isFieldMandatory($field)) {
-                $mandatoryFields[] = $field['name'];
-            }
-
-            if (isset($field['attributes'])) {
-                $attributes = $field['attributes'];
-
-                $mandatoryFields = $this->getMandatoryFields($attributes, $mandatoryFields);
-            }
-
-            if (!empty($mandatoryFields)) {
-                $mandatoryElements[$field['name']] = $mandatoryFields;
-            }
-        }
-
-        return $mandatoryElements;
-    }
-
-    /**
-     * Returns mandatory attributes.
-     *
-     * @param $subElement
-     *
-     * @return array
-     */
-    public function getMandatoryAttributes($subElement): array
-    {
-        return array_key_exists('attributes', $subElement) ? $this->mandatoryAttributes($subElement['attributes']) : [];
-    }
-
-    /**
-     * Returns mandatory sub elements.
-     *
-     * @param $subElement
-     *
-     * @return array
-     */
-    public function getMandatorySubElements($subElement): array
-    {
-        return array_key_exists('sub_elements', $subElement) ? $this->mandatorySubElements($subElement['sub_elements']) : [];
-    }
 
     /**
      * Checks if attribute is complete.
@@ -146,212 +51,7 @@ class ElementCompleteService
 
         $elementSchema = getElementSchema($this->element);
 
-        foreach ($mandatoryAttributes as $mandatoryAttribute) {
-            if (array_key_exists('dependent_attributes', $elementSchema) && array_key_exists($mandatoryAttribute, $elementSchema['dependent_attributes'])) {
-                $parentLevel = $elementSchema['attributes'];
-
-                if (
-                    array_key_exists('sub_element', $elementSchema['dependent_attributes'][$mandatoryAttribute])
-                    && !empty($elementSchema['dependent_attributes'][$mandatoryAttribute]['sub_element'])
-                ) {
-                    $parentLevel = $elementSchema['sub_elements'][$elementSchema['dependent_attributes'][$mandatoryAttribute]['sub_element']]['attributes'];
-                }
-
-                $parent = $parentLevel[$mandatoryAttribute]['parent'];
-
-                /*checks if parent attribute have specific value for child attribute to be relevant*/
-                if (!in_array($data[$parent['name']], $parent['value'], true)) {
-                    continue;
-                }
-            }
-
-            if (!array_key_exists($mandatoryAttribute, $data) || (empty($data[$mandatoryAttribute]))) {
-                //dd('isAttributeDataCompleted fx called1', ' Attribute is empty', 'attribute-check:', $mandatoryAttributes, $mandatoryAttribute, $data);
-
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Checks if single dimension attribute is complete.
-     *
-     * @param $elementSchema
-     * @param $data
-     *
-     * @return bool
-     * @throws \JsonException
-     */
-    public function isSingleDimensionAttributeCompleted($elementSchema, $data): bool
-    {
-        $mandatoryAttributes = $this->getMandatoryAttributes($elementSchema);
-
-        if (!empty($mandatoryAttributes)) {
-            if (empty($data)) {
-                return false;
-            }
-
-            if (!$this->isAttributeDataCompleted($mandatoryAttributes, $data)) {
-                //dd('singleDimensionAttributeCheck fx called1', 'Level2 single dimension attribute is empty', 'attribute-check:', $mandatoryAttributes, $data);
-
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Checks if multi dimension attribute is complete.
-     *
-     * @param $elementSchema
-     * @param $data
-     *
-     * @return bool
-     * @throws \JsonException
-     */
-    public function isMultiDimensionAttributeCompleted($elementSchema, $data): bool
-    {
-        $mandatoryAttributes = $this->getMandatoryAttributes($elementSchema);
-
-        if (!empty($mandatoryAttributes)) {
-            if (empty($data)) {
-                return false;
-            }
-
-            foreach ($data as $datum) {
-                if (!$this->isAttributeDataCompleted($mandatoryAttributes, $datum)) {
-                    return false;
-                }
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Checks if sub element is complete.
-     *
-     * @param $mandatorySubElements
-     * @param $data
-     *
-     * @return bool
-     */
-    public function isSubElementDataCompleted($mandatorySubElements, $data): bool
-    {
-        if (is_variable_null($data)) {
-            return false;
-        }
-
-        if (empty($mandatorySubElements)) {
-            return true;
-        }
-
-        if (empty($data)) {
-            return false;
-        }
-
-        foreach ($mandatorySubElements as $key => $mandatorySubElement) {
-            if (!array_key_exists($key, $data)) {
-                //dd('isSubElementDataCompleted fx called1', 'Whole Sub element has not filled yet', 'sub-element-check:', $mandatorySubElement, $data);
-
-                return false;
-            }
-            $items = $data[$key];
-
-            if (empty($items)) {
-                //dd('isSubElementDataCompleted fx called2', 'Sub element array is empty', 'sub-element-check:', $mandatorySubElement, $data, $items);
-
-                return false;
-            }
-
-            foreach ($mandatorySubElement as $mandatoryField) {
-                foreach ($items as $item) {
-                    if (!array_key_exists($mandatoryField, $item) || (empty($item[$mandatoryField]))) {
-                        //dd('isSubElementDataCompleted fx called3', 'Sub element is empty', 'sub-element-check:', $mandatoryField, $item);
-
-                        return false;
-                    }
-                }
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Checks if element's attribute and sub elements both are complete.
-     *
-     * @param $mandatoryAttributes
-     * @param $mandatorySubElements
-     * @param $data
-     *
-     * @return bool
-     * @throws \JsonException
-     */
-    public function isElementCompleted($mandatoryAttributes, $mandatorySubElements, $data): bool
-    {
-        if (!empty($mandatoryAttributes) || !empty($mandatorySubElements)) {
-            if (empty($data)) {
-                return false;
-            }
-
-            foreach ($data as $datum) {
-                if (!$this->isAttributeDataCompleted($mandatoryAttributes, $datum)) {
-                    //dd('isElementCompleted fx is called1', 'Attribute is empty', 'mandatory-attributes:', $mandatoryAttributes, $data, $datum);
-
-                    return false;
-                }
-
-                if (!$this->isSubElementDataCompleted($mandatorySubElements, $datum)) {
-                    //dd('isElementCompleted fx is called2', 'Sub element is empty', 'mandatory-sub-element:', $mandatorySubElements, $data, $datum);
-
-                    return false;
-                }
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * Checks if sub element with attributes is complete.
-     *
-     * @param $subElements
-     * @param $data
-     *
-     * @return bool
-     * @throws \JsonException
-     */
-    public function isSubElementCompleted($subElements, $data): bool
-    {
-        foreach ($subElements as $key => $subElement) {
-            $mandatorySubElementAttributes = $this->getMandatoryAttributes($subElement);
-            $mandatoryChildSubElements = $this->getMandatorySubElements($subElement);
-
-            if (!empty($mandatorySubElementAttributes) || !empty($mandatoryChildSubElements)) {
-                if (!array_key_exists($key, $data)) {
-                    //dd('isSubElementCompleted fx is called1', 'Sub element key not present', 'mandatory-sub-element_attributes:', $mandatorySubElementAttributes, 'mandatory-child-sub-elements: ',
-                    //  $mandatoryChildSubElements, $key, $data);
-
-                    return false;
-                }
-
-                if (empty($data[$key])) {
-                    //dd('isSubElementCompleted fx is called2', 'Sub element empty', 'mandatory-sub-element_attributes:', $mandatorySubElementAttributes, 'mandatory-child-sub-elements: ',
-                    //$mandatoryChildSubElements, $key, $data);
-
-                    return false;
-                }
-                if (!$this->isElementCompleted($mandatorySubElementAttributes, $mandatoryChildSubElements, $data[$key])) {
-                    return false;
-                }
-            }
-        }
-
-        return true;
+        return $this->checkAttributeDataStatus($mandatoryAttributes, $data, $elementSchema);
     }
 
     /**
@@ -365,20 +65,6 @@ class ElementCompleteService
     public function singleDimensionAttributeCheck($data): bool
     {
         return $this->isSingleDimensionAttributeCompleted(getElementSchema($this->element), $data);
-    }
-
-    /**
-     * Checks if level one attribute and sub element is complete.
-     *
-     * @param $elementSchema
-     * @param $data
-     *
-     * @return bool
-     * @throws \JsonException
-     */
-    public function isLevelOneMultiDimensionDataCompleted($elementSchema, $data): bool
-    {
-        return $this->isElementCompleted($this->getMandatoryAttributes($elementSchema), $this->getMandatorySubElements($elementSchema), $data);
     }
 
     /**
@@ -415,55 +101,6 @@ class ElementCompleteService
         $elementSchema = getElementSchema($this->element);
 
         return $this->isSubElementCompleted($elementSchema['sub_elements'], $data);
-    }
-
-    /**
-     * Checks if level two attribute and sub element is complete.
-     *
-     * @param $elementSchema
-     * @param $data
-     *
-     * @return bool
-     * @throws \JsonException
-     */
-    public function isLevelTwoMultiDimensionDataCompleted($elementSchema, $data): bool
-    {
-        if (!$this->isMultiDimensionAttributeCompleted($elementSchema, $data)) {
-            return false;
-        }
-
-        $subElements = getArr($elementSchema, 'sub_elements');
-        $mandatorySubElementsFlag = false;
-
-        foreach ($subElements as $subElement) {
-            $mandatorySubElementAttributes = $this->getMandatoryAttributes($subElement);
-
-            if (!empty($mandatorySubElementAttributes)) {
-                $mandatorySubElementsFlag = true;
-                break;
-            }
-
-            $mandatoryChildSubElements = $this->getMandatorySubElements($subElement);
-
-            if (!empty($mandatoryChildSubElements)) {
-                $mandatorySubElementsFlag = true;
-                break;
-            }
-        }
-
-        if ($mandatorySubElementsFlag) {
-            if (empty($data)) {
-                return false;
-            }
-
-            foreach ($data as $datum) {
-                if (!$this->isSubElementCompleted($subElements, $datum)) {
-                    return false;
-                }
-            }
-        }
-
-        return true;
     }
 
     /**
@@ -507,15 +144,10 @@ class ElementCompleteService
                 continue;
             }
 
-            if (!array_key_exists($key, $data)) {
-                //dd('isLevelThreeSingleDimensionElementCompleted fx called1', 'sub-element-empty', 'sub-element-check:', $mandatorySubElementAttributes, $key, $data);
-
-                return false;
-            }
-
-            if (empty($data[$key])) {
-                //dd('isLevelThreeSingleDimensionElementCompleted fx called2', 'sub-element-empty', 'sub-element-check:', $mandatorySubElementAttributes, $key, $data);
-
+            if (
+                !array_key_exists($key, $data)
+                || empty($data[$key])
+            ) {
                 return false;
             }
 
@@ -526,8 +158,6 @@ class ElementCompleteService
                     return false;
                 }
             }
-
-            $tempData = $data[$key];
 
             foreach ($tempData as $tempDatum) {
                 if (!$this->isSubElementCompleted($subElement['sub_elements'], $tempDatum)) {
@@ -1029,11 +659,10 @@ class ElementCompleteService
     public function isBaselineCompleted($data, $subElement): bool
     {
         foreach ($data as $baselineDatum) {
-            if (!$this->isLevelOneMultiDimensionDataCompleted($subElement['sub_elements']['comment'], getArr($baselineDatum, 'comment'))) {
-                return false;
-            }
-
-            if (!$this->isLevelTwoMultiDimensionDataCompleted($subElement['sub_elements']['document_link'], getArr($baselineDatum, 'document_link'))) {
+            if (
+                !$this->isLevelOneMultiDimensionDataCompleted($subElement['sub_elements']['comment'], getArr($baselineDatum, 'comment'))
+                || !$this->isLevelTwoMultiDimensionDataCompleted($subElement['sub_elements']['document_link'], getArr($baselineDatum, 'document_link'))
+            ) {
                 return false;
             }
         }
@@ -1052,7 +681,10 @@ class ElementCompleteService
      */
     public function isResultAndIndicatorElementCompleted($elementSchema, $data): bool
     {
-        if (!$this->isSingleDimensionAttributeCompleted($elementSchema, $data)) {
+        if (
+            !$this->isSingleDimensionAttributeCompleted($elementSchema, $data)
+            || !$this->isLevelTwoMultiDimensionDataCompleted($elementSchema['sub_elements']['document_link'], getArr($data, 'document_link'))
+        ) {
             return false;
         }
 
@@ -1060,10 +692,6 @@ class ElementCompleteService
             if (!$this->isLevelOneMultiDimensionDataCompleted($elementSchema['sub_elements'][$item], getArr($data, $item))) {
                 return false;
             }
-        }
-
-        if (!$this->isLevelTwoMultiDimensionDataCompleted($elementSchema['sub_elements']['document_link'], getArr($data, 'document_link'))) {
-            return false;
         }
 
         return true;
@@ -1083,11 +711,10 @@ class ElementCompleteService
         $elementSchema = getElementSchema($this->element);
 
         foreach ($data as $datum) {
-            if (!$this->isResultAndIndicatorElementCompleted($elementSchema, $datum)) {
-                return false;
-            }
-
-            if (!$this->isBaselineCompleted($datum['baseline'], $elementSchema['sub_elements']['baseline'])) {
+            if (
+                !$this->isResultAndIndicatorElementCompleted($elementSchema, $datum)
+                || !$this->isBaselineCompleted($datum['baseline'], $elementSchema['sub_elements']['baseline'])
+            ) {
                 return false;
             }
         }
@@ -1166,15 +793,11 @@ class ElementCompleteService
     {
         [$resultData, $indicatorData, $periodData] = $this->getFormattedResults($activity);
 
-        if (is_variable_null($periodData) || !$this->isPeriodElementCompleted($periodData)) {
-            return false;
-        }
-
-        if (is_variable_null($indicatorData) || !$this->isIndicatorElementCompleted($indicatorData)) {
-            return false;
-        }
-
-        if (is_variable_null($resultData) || !$this->isResultElementDataCompleted($resultData)) {
+        if (
+            (is_variable_null($periodData) || !$this->isPeriodElementCompleted($periodData))
+            || (is_variable_null($indicatorData) || !$this->isIndicatorElementCompleted($indicatorData))
+            || (is_variable_null($resultData) || !$this->isResultElementDataCompleted($resultData))
+        ) {
             return false;
         }
 
@@ -1250,24 +873,45 @@ class ElementCompleteService
                     $this->setDefaultValues($datum, $activity);
                 }
 
-                if ($key === 'narrative') {
-                    $this->tempNarrative = $datum;
-                }
-
-                if ($key === 'amount') {
-                    $this->tempAmount = $datum;
-                }
-
-                if ($key === 'language' && empty($datum) && !empty($this->tempNarrative)) {
-                    $data['language'] = Arr::get($activity->default_field_values, 'default_language', null);
-                }
-
-                if ($key === 'currency' && empty($datum) && !empty($this->tempAmount)) {
-                    $data['currency'] = Arr::get($activity->default_field_values, 'default_currency', null);
-                }
+                $this->setTempNarrative((string) $key, $datum);
+                $this->setTempAmount((string) $key, $datum);
+                $this->updateLanguage($data, (string) $key, $datum, $activity);
+                $this->updateCurrency($data, (string) $key, $datum, $activity);
             }
         }
 
         return $data;
+    }
+
+    /**
+     * Assigns language to $data['language'] variable.
+     *
+     * @param array $data
+     * @param string $key
+     * @param $datum
+     * @param $activity
+     * @return void
+     */
+    public function updateLanguage(array &$data, string $key, $datum, $activity): void
+    {
+        if ($key === 'language' && empty($datum) && !empty($this->tempNarrative)) {
+            $data['language'] = Arr::get($activity->default_field_values, 'default_language', null);
+        }
+    }
+
+    /**
+     * Assigns currency to $data['currency'] variable.
+     *
+     * @param array $data
+     * @param string $key
+     * @param $datum
+     * @param $activity
+     * @return void
+     */
+    public function updateCurrency(array &$data, string $key, $datum, $activity): void
+    {
+        if ($key === 'currency' && empty($datum) && !empty($this->tempAmount)) {
+            $data['currency'] = Arr::get($activity->default_field_values, 'default_currency', null);
+        }
     }
 }
